+++ templates/components/vectordbs/typescript/astra/generate.ts
@@ -1,7 +1,7 @@
 /* eslint-disable turbo/no-undeclared-env-vars */
 import * as dotenv from "dotenv";
 import { VectorStoreIndex, storageContextFromDefaults } from "llamaindex";
-import { AstraDBVectorStore } from "llamaindex/storage/vectorStore/AstraDBVectorStore";
+import { AstraDBVectorStore } from "llamaindex/vector-store/AstraDBVectorStore";
 import { getDocuments } from "./loader";
 import { initSettings } from "./settings";
 import { checkRequiredEnvVars } from "./shared";
@@ -15,13 +15,12 @@ async function loadAndIndex() {
   // create vector store and a collection
   const collectionName = process.env.ASTRA_DB_COLLECTION!;
   const vectorStore = new AstraDBVectorStore();
-  await vectorStore.create(collectionName, {
+  await vectorStore.createAndConnect(collectionName, {
     vector: {
       dimension: parseInt(process.env.EMBEDDING_DIM!),
       metric: "cosine",
     },
   });
-  await vectorStore.connect(collectionName);
 
   // create index from documents and store them in Astra
   console.log("Start creating embeddings...");